+++ wxc_sdk/rest.py
@@ -377,14 +377,14 @@ class RestSession(Session):
         """
         return self._rest_request('PUT', *args, **kwargs)
 
-    def rest_delete(self, *args, **kwargs) -> None:
+    def rest_delete(self, *args, **kwargs) -> StrOrDict:
         """
         DELETE request
 
         :param args:
         :param kwargs:
         """
-        self._rest_request('DELETE', *args, **kwargs)
+        return self._rest_request('DELETE', *args, **kwargs)
 
     def rest_patch(self, *args, **kwargs) -> StrOrDict:
         """